% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{estimate.km}
\alias{estimate.km}
\title{Kaplan-Meier estimate of the survivor function}
\usage{
estimate.km(y, delta, adjust = FALSE)
}
\arguments{
\item{y}{A numeric vector (length \code{n}) of the observed event time.}

\item{delta}{A numeric vector (length \code{n}) of the censoring indicator.}

\item{adjust}{A logical to indicate if imposing an adjustment in computing
the hazard components to avoid the KME being zero when
time greater than the largest observation time.}
}
\value{
A \code{\link[stats]{stepfun}} of the estimate.
}
\description{
Kaplan-Meier estimate of the survivor function
}
\keyword{internal}
