% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract.km}
\alias{extract.km}
\title{Extract probability masses from Kaplan-Meier estimate}
\usage{
extract.km(km)
}
\arguments{
\item{km}{A \code{\link[stats]{stepfun}} of the Kaplan-Meier estimate
computed from \code{\link{estimate.km}}.}
}
\value{
A list of the extract masses and the corresponding time
}
\description{
Extract probability masses from Kaplan-Meier estimate
}
\keyword{internal}
