% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gren.R
\name{f.hat}
\alias{f.hat}
\title{Evaluate the Grenander density estimate at \code{x}}
\usage{
f.hat(x, gren.times, gren.val)
}
\arguments{
\item{x}{A numeric vector for which the value of Grenander density estimate is computed.}

\item{gren.times}{A numeric vector of points at which there is a jump.}

\item{gren.val}{A numeric vector of Grenander density estimate corresponds to `gren.times`.}
}
\value{
A numeric vector of the computed values
}
\description{
Evaluate the Grenander density estimate at \code{x}
}
\keyword{internal}
