% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{grenander.crit.val}
\alias{grenander.crit.val}
\title{Compute critical value for the test using \eqn{\hat{f}^{G}}}
\usage{
grenander.crit.val(gren, a, tau, c.shift = 1, eps = 0.01, alpha = 0.05)
}
\arguments{
\item{gren}{A list from \code{\link{grenander.test}}.}

\item{tau}{A numeric of the parameter \eqn{\tau} for the test.}

\item{c.shift}{A positive number of the constant \eqn{c}.}

\item{eps}{A numeric of the \eqn{\epsilon} for the test.}

\item{alpha}{A numeric of the significance level.}
}
\value{
A numeric of the computed critical value
}
\description{
Compute critical value:
\deqn{
 \frac{\epsilon\hat{F}_{n}(\tau_{G})}{\tau - \tau_{G}} - A_{1}^{-1}n^{-(1-a)/2}Q_{1 - \alpha}^{G}
}
where \eqn{\hat{F}_{n}} is the KME,
\eqn{A_1=\sqrt{c[1-G(\tau_G-)]/f(\tau_G)}}, and
\eqn{Q_{\alpha}^{G}} is the \eqn{\alpha}-quantile of the distribution \eqn{D_R[W(t)](1)},
with \eqn{W(t)} denotes a Brownian motion and
\eqn{D_R[Z(t)](b)} is the right derivative of the LCM on
\eqn{[0,\infty)} of the process \eqn{Z(t)} at the point \eqn{t=b}.
}
\keyword{internal}
