% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{grenander.p.val}
\alias{grenander.p.val}
\title{Compute \eqn{p}-value for the test using \eqn{\hat{f}^{G}}}
\usage{
grenander.p.val(gren, a, tau, c.shift = 1, eps = 0.01)
}
\arguments{
\item{gren}{A list from \code{\link{grenander.test}}.}

\item{tau}{A numeric of the parameter \eqn{\tau} for the test.}

\item{c.shift}{A positive number of the constant \eqn{c}.}

\item{eps}{A numeric of the \eqn{\epsilon} for the test.}
}
\value{
A numeric of the computed \eqn{p}-value
}
\description{
Compute \eqn{p}-value for the test using \eqn{\hat{f}^{G}}
}
\keyword{internal}
