% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{grenander.stat}
\alias{grenander.stat}
\title{Compute the test statistic for the test using \eqn{\hat{f}^{G}}}
\usage{
grenander.stat(gren, a, c.shift = 1, eps = 0.01)
}
\arguments{
\item{gren}{A list from \code{\link{grenander.test}}.}

\item{a}{A numeric of the constant \eqn{a\in(1/3,1)} controls
the rate of convergence of the estimator.}

\item{c.shift}{A positive number of the constant \eqn{c}.}

\item{eps}{A numeric of the \eqn{\epsilon} for the test.}
}
\value{
A numeric of the computed test statistic.
}
\description{
Compute the test statistic for the test using \eqn{\hat{f}^{G}}:
\deqn{\hat{f}_{n}^{G}(\tau_{G} - cn^{-a})}.
}
\keyword{internal}
