% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothing.R
\name{i0}
\alias{i0}
\title{Coefficients for the boundary kernel}
\usage{
i0(s)
}
\arguments{
\item{s}{A numeric vector or matrix of points for which the value of
the above is computed.}
}
\value{
A numeric vector or matrix of the computed values.
}
\description{
Compute the following:
\deqn{\int_{-\infty}^s k(u)du,}
where \eqn{k(u) = \frac{35}{32}(1-u^2)^3I_{[-1,1]}(u)}.
}
\keyword{internal}
