% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothing.R
\name{integrated.kern}
\alias{integrated.kern}
\title{Integrated tri-weight kernel function}
\usage{
integrated.kern(x)
}
\arguments{
\item{x}{A numeric vector or matrix of points for which the value of
\code{integrated.kern} is computed.}
}
\value{
A numeric vector or matrix of the computed values.
}
\description{
Integrated tri-weight kernel function:
\deqn{\int_{-\infty}^x k(u)du,}
where \eqn{k(u) = \frac{35}{32}(1-u^2)^3I_{[-1,1]}(u)}.
}
\keyword{internal}
