% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothing.R
\name{kern}
\alias{kern}
\title{Tri-weight kernel function}
\usage{
kern(x)
}
\arguments{
\item{x}{A numeric vector or matrix of points for which the value of
tri-weight kernel is computed.}
}
\value{
A numeric vector or matrix of the computed values.
}
\description{
Tri-weight kernel function:
\eqn{k(u) = \frac{35}{32}(1-u^2)^3I_{[-1,1]}(u)}.
}
\keyword{internal}
