% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothing.R
\name{phi}
\alias{phi}
\title{Coefficients for the boundary kernel}
\usage{
phi(x)
}
\arguments{
\item{x}{A numeric vector or matrix of points for which the value of
the above is computed.}
}
\value{
A numeric vector or matrix of the computed values.
}
\description{
Compute the coefficients \eqn{\phi(x)} for \eqn{k(u)}.
}
\keyword{internal}
