% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfu.R
\name{print.sfu.cov}
\alias{print.sfu.cov}
\title{Print a sfu.cov object}
\usage{
\method{print}{sfu.cov}(x, ...)
}
\arguments{
\item{x}{An \code{sfu.cov} object.}

\item{\dots}{Additional print arguments.}
}
\value{
A printout mentioned in Details.
}
\description{
Print a summary of the sufficient follow-up with categorical covariates test result.
Let \eqn{F_{u}(\cdot|x)} be the conditional distribution of
the survival time for the uncured subject given \eqn{X=x} and
\eqn{G(\cdot|x)} be the conditional distribution of
the random right censoring time given \eqn{X=x}.
The test hypothesis is:
\deqn{
   {H}_{0}: q_{1 - \epsilon}(x) \geq \tau_{G}(x)\text{ for some }x
   \quad\text{versus}\quad
   {H}_{1}: q_{1 - \epsilon}(x) < \tau_{G}(x)\text{ for all }x,
}
where \eqn{q_{1 - \epsilon}(x) = \inf\lbrace{t \geq 0}:{F_{u}(t|x) \geq 1 - \epsilon}\rbrace} and
\eqn{\tau_{G}(x)=\sup\lbrace{t\geq0}:{G(t|x)<1}\rbrace}.
The test hypothesis for fixed covariate value \eqn{x} is:
\deqn{
   {H}_{0x}: q_{1 - \epsilon}(x) \geq \tau_{G}(x)
   \quad\text{versus}\quad
   {H}_{1x}: q_{1 - \epsilon}(x) < \tau_{G}(x),
}
}
\details{
The value of the test statistic, p-value of the test \eqn{H_{0x}},
the estimate \eqn{\hat{x}_{\ast,n}} and the setting are printed.
}
