% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfu.R
\name{print.sfu}
\alias{print.sfu}
\title{Print a sfu object}
\usage{
\method{print}{sfu}(x, ...)
}
\arguments{
\item{x}{An \code{sfu} object.}

\item{\dots}{Additional print arguments.}
}
\value{
A printout mentioned in Details.
}
\description{
Print a summary of the sufficient follow-up test result.
}
\details{
The value of the test statistic, p-value of the test, and the setting are printed.
}
