% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothing.R
\name{psi}
\alias{psi}
\title{Coefficients for the boundary kernel}
\usage{
psi(x)
}
\arguments{
\item{x}{A numeric vector or matrix of points for which the value of
the above is computed.}
}
\value{
A numeric vector or matrix of the computed values.
}
\description{
Compute the coefficients \eqn{\psi(x)} for \eqn{uk(u)}.
}
\keyword{internal}
