% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{resample.censor}
\alias{resample.censor}
\title{Generate bootstrap sample of censoring time}
\usage{
resample.censor(n, rev.km)
}
\arguments{
\item{n}{An integer of sample size.}

\item{rev.km}{A \code{\link[stats]{stepfun}} of the Kaplan-Meier estimate
for the censoring distribution computed from \code{\link{estimate.km}}.}
}
\value{
A numeric vector of bootstrap sample of censoring time.
}
\description{
Generate bootstrap sample of censoring time
}
\keyword{internal}
