% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{resample.survival}
\alias{resample.survival}
\title{Generate bootstrap sample of event time}
\usage{
resample.survival(n, skm.df, tau.c)
}
\arguments{
\item{n}{An integer of sample size.}

\item{skm.df}{A \code{data.frame} of the Smoothed LCM of KME
computed from \code{\link{smooth.kme}}.}

\item{tau.c}{A numeric of the end of the study time.}
}
\value{
A numeric vector of bootstrap sample of event time.
}
\description{
Generate bootstrap sample of event time
}
\keyword{internal}
