% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfu.R
\name{sfu.cov.test}
\alias{sfu.cov.test}
\title{Test for sufficient follow-up in survival data with a cure fraction
and categorical covariates
using shape constrained density estimator}
\usage{
sfu.cov.test(
  y,
  delta,
  X,
  tau,
  eps = 0.01,
  alpha = 0.05,
  gamma = 0.025,
  n.boot = 1000L
)
}
\arguments{
\item{y}{a numeric vector of the observed survival times, \eqn{Y}.}

\item{delta}{a numeric vector of the censoring indicators, \eqn{\Delta}.}

\item{X}{a numeric vector of the categorical covariates, \eqn{X}.}

\item{tau}{a numeric of the parameter \eqn{\tau} for the test.
It must be greater than the observed survival time \code{max(y)}. See 'Details' below.}

\item{eps}{a numeric in \eqn{(0, 1)} of the \eqn{\epsilon} for the test.
Default is \code{0.01}.}

\item{alpha}{A numeric in \eqn{(0, 1)} of the significance level for the test.
Default is \code{0.05}.}

\item{gamma}{A numeric in \eqn{(0, 1)} of the parameter used to estimate \eqn{x_\ast}.}

\item{n.boot}{An integer of number of bootstrap samples. Default is \code{1000L}.}
}
\value{
A list with class \code{sfu.cov} containing the following components:
\describe{
\item{\code{statistic}}{the value of the test statistic for \eqn{H_{0x}}.}
\item{\code{p.value}}{the \eqn{p}-value of the test for \eqn{H_{0x}}.}
\item{\code{crit.val}}{the critical value of the test for \eqn{H_{0x}} using the significance level \code{alpha}.}
\item{\code{alpha}}{the value of the significance level used.}
\item{\code{y.max.x}}{the value of \eqn{Y_{(n)}^x}.}
\item{\code{x.ast.hat}}{the value of \eqn{\hat{x}_{\ast,n}}.}
\item{\code{tau}}{the value of \code{tau} used.}
\item{\code{eps}}{the value of \code{eps} used.}
\item{\code{x.unique}}{the unique value of \code{X}.}
\item{\code{data.name}}{a character string giving the names of the data.}
}
}
\description{
Test for sufficient follow-up in survival data with a cure fraction
and categorical covariates \eqn{X}
using the smoothed Grenander estimator \eqn{\hat{f}_{nh}^{SG}}.
Let \eqn{F_{u}(\cdot|x)} be the conditional distribution of
the survival time for the uncured subject given \eqn{X=x} and
\eqn{G(\cdot|x)} be the conditional distribution of
the random right censoring time given \eqn{X=x}.
The test hypothesis is:
\deqn{
   {H}_{0}: q_{1 - \epsilon}(x) \geq \tau_{G}(x)\text{ for some }x
   \quad\text{versus}\quad
   {H}_{1}: q_{1 - \epsilon}(x) < \tau_{G}(x)\text{ for all }x,
}
where \eqn{q_{1 - \epsilon}(x) = \inf\lbrace{t \geq 0}:{F_{u}(t|x) \geq 1 - \epsilon}\rbrace} and
\eqn{\tau_{G}(x)=\sup\lbrace{t\geq0}:{G(t|x)<1}\rbrace}.
The test hypothesis for fixed covariate value \eqn{x} is:
\deqn{
   {H}_{0x}: q_{1 - \epsilon}(x) \geq \tau_{G}(x)
   \quad\text{versus}\quad
   {H}_{1x}: q_{1 - \epsilon}(x) < \tau_{G}(x),
}
}
\details{
The choice of \eqn{\tau} can be based on some prior knowledge, for example,
 one can take \eqn{\tau > \tau_{G}(x)} such that it is almost impossible
 for the event to happen after \eqn{\tau} given \eqn{X = x} for all \eqn{x}.
}
