% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfu.R
\name{sfu.test}
\alias{sfu.test}
\title{Test for sufficient follow-up in survival data with a cure fraction
using shape constrained density estimator}
\usage{
sfu.test(
  y,
  delta,
  tau,
  tau.g = NULL,
  eps = 0.01,
  method = c("g", "sg"),
  alpha = 0.05,
  n.boot = 1000L
)
}
\arguments{
\item{y}{a numeric vector of the observed survival times, \eqn{Y_i}.}

\item{delta}{a numeric vector of the censoring indicators, \eqn{\Delta_i}.}

\item{tau}{a numeric of the parameter \eqn{\tau} for the test.
It must be greater than \code{tau.g}. See 'Details' below.}

\item{tau.g}{a numeric of the end of the study time, \eqn{\tau_{G}}.
See 'Details' for the meaning of \code{NULL} (the default).}

\item{eps}{a numeric in \eqn{(0, 1)} of the \eqn{\epsilon} for the test.
Default is \code{0.01}.}

\item{method}{a character string indicating which shape constrained density
estimator is to be used for the test.
One of "g" (indicating Grenander estimator \eqn{\hat{f}_{n}^{G}}), or
"sg" (smoothed Grenander estimator \eqn{\hat{f}_{nh}^{SG}}).}

\item{alpha}{A numeric in \eqn{(0, 1)} of the significance level for the test.
Default is \code{0.05}.}

\item{n.boot}{An integer of number of bootstrap samples when
the smoothed Grenander estimator \eqn{\hat{f}_{nh}^{SG}} is used, i.e. \code{method = "sg"}.
Default is \code{1000L}.}
}
\value{
A list with class \code{sfu} containing the following components:
\describe{
\item{\code{statistic}}{the value of the test statistic.}
\item{\code{p.value}}{the \eqn{p}-value of the test.}
\item{\code{crit.val}}{the critical value of the test using the significance level \code{alpha}.}
\item{\code{method}}{a character string indicating how the density was estimated.}
\item{\code{alpha}}{the value of the significance level used.}
\item{\code{tau.g}}{the value of \code{tau.g} used.}
\item{\code{tau}}{the value of \code{tau} used.}
\item{\code{eps}}{the value of \code{eps} used.}
\item{\code{data.name}}{a character string giving the names of the data.}
}
}
\description{
Test for sufficient follow-up in survival data with a cure fraction
using the smoothed Grenander estimator \eqn{\hat{f}_{nh}^{SG}} or
the Grenander estimator \eqn{\hat{f}_{n}^{G}}.
Let \eqn{F_{u}} be the distribution of the survival time for the uncured subject and
\eqn{G} be the distribution of the random right censoring time.
The test hypothesis is:
\deqn{
   \tilde{H}_{0}: q_{1 - \epsilon} \geq \tau_{G}
   \quad\text{versus}\quad
   \tilde{H}_{a}: q_{1 - \epsilon} < \tau_{G},
}
where \eqn{q_{1 - \epsilon} = \inf\lbrace{t \geq 0}:{F_{u}(t) \geq 1 - \epsilon}\rbrace} and
\eqn{\tau_{G}=\sup\lbrace{t\geq0}:{G(t)<1}\rbrace}.
}
\details{
If \code{tau.g} is \code{NULL},
 the maximum observed survival time \code{max(y)} is used for the test.
 The choice of \eqn{\tau} can be based on some prior knowledge, for example,
 one can take \eqn{\tau > \tau_{G}} such that it is almost impossible
 for the event to happen after \eqn{\tau}.
}
