% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{smooth.grenander.bootstrap}
\alias{smooth.grenander.bootstrap}
\title{Generate bootstrap sample of smooth Grenander density estimate}
\usage{
smooth.grenander.bootstrap(rev.km, gcmlcm.fit, skm.df, tau.c, n)
}
\arguments{
\item{rev.km}{A \code{\link[stats]{stepfun}} of the Kaplan-Meier estimate
for the censoring distribution computed from \code{\link{estimate.km}}.}

\item{gcmlcm.fit}{A list of LCM of KME returned from \code{\link[fdrtool]{gcmlcm}}.}

\item{skm.df}{A \code{data.frame} of the Smoothed LCM of KME
computed from \code{\link{smooth.kme}}.}

\item{tau.c}{A numeric of the end of the study time.}

\item{n}{An integer of sample size.}
}
\value{
A list containing bootstrap samples, including \eqn{\hat{f}_{nh}^{SG^\ast}(\tau_{G})}.
}
\description{
Generate bootstrap sample of smooth Grenander density estimate
}
\keyword{internal}
