% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{smooth.grenander.bootstrap.undersmooth}
\alias{smooth.grenander.bootstrap.undersmooth}
\title{Generate bootstrap sample of smooth Grenander density estimate
with undersmoothing}
\usage{
smooth.grenander.bootstrap.undersmooth(
  rev.km,
  gcmlcm.fit,
  tau.c,
  n,
  skm.bw.multi
)
}
\arguments{
\item{rev.km}{A \code{\link[stats]{stepfun}} of the Kaplan-Meier estimate
for the censoring distribution computed from \code{\link{estimate.km}}.}

\item{gcmlcm.fit}{A list of LCM of KME returned from \code{\link[fdrtool]{gcmlcm}}.}

\item{tau.c}{A numeric of the end of the study time.}

\item{n}{An integer of sample size.}

\item{skm.bw.multi}{A numeric of scaling to the bandwidth for smoothing
the density estimator for generating bootstrap samples.}
}
\value{
A list containing bootstrap samples, including \eqn{\hat{f}_{nh}^{SG^\ast}(Y_{(n)})}.
}
\description{
Generate bootstrap sample of smooth Grenander density estimate
with undersmoothing
}
\keyword{internal}
