% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{smooth.grenander.crit.val}
\alias{smooth.grenander.crit.val}
\title{Compute critical value from the bootstrap samples for the test using \eqn{\hat{f}_{nh}^{SG}}}
\usage{
smooth.grenander.crit.val(sg, tau, tau.c, eps = 0.01, alpha = 0.05)
}
\arguments{
\item{sg}{A list from \code{\link{smooth.grenander.test}}.}

\item{tau}{A numeric of the parameter \eqn{\tau} for the test.}

\item{tau.c}{A numeric of the end of the study time.}

\item{eps}{A numeric of the \eqn{\epsilon} for the test.}

\item{alpha}{A numeric of the significance level.}
}
\value{
A numeric of the computed critical value
}
\description{
Compute critical value from the bootstrap samples:
\deqn{
 \frac{\epsilon \hat{F}_{n}(\tau_{G})}{\tau - \tau_{G}} + n^{-2/5}Q_{\alpha}^{SG},
}
where \eqn{\hat{F}_{n}} is the KME,
\eqn{Q_{\alpha}^{SG}} is the \eqn{\alpha}-quantile of the bootstrap estimates
\eqn{n^{2/5}\{\hat{f}_{nh}^{SG^{\ast}}(\tau_{G})-\tilde{f}_{nh_{0}}(\tau_{G})\}},
with \eqn{\tilde{f}_{nh_{0}}} being the derivative of the smoothed LCM of KME from
\code{\link{smooth.kme}}.
}
\keyword{internal}
