% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{smooth.grenander.p.val}
\alias{smooth.grenander.p.val}
\title{Compute \eqn{p}-value from the bootstrap samples for the test \eqn{\hat{f}_{nh}^{SG}}}
\usage{
smooth.grenander.p.val(sg, tau, tau.c, eps = 0.01)
}
\arguments{
\item{sg}{A list from \code{\link{smooth.grenander.test}}.}

\item{tau}{A numeric of the parameter \eqn{\tau} for the test.}

\item{tau.c}{A numeric of the end of the study time.}

\item{eps}{A numeric of the \eqn{\epsilon} for the test.}
}
\value{
A numeric of the computed \eqn{p}-value
}
\description{
Compute \eqn{p}-value from the bootstrap samples for the test \eqn{\hat{f}_{nh}^{SG}}
}
\keyword{internal}
