% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{smooth.grenander.test}
\alias{smooth.grenander.test}
\title{Testing sufficient follow-up using Smoothed Grenander Estimator \eqn{\hat{f}_{nh}^{SG}}}
\usage{
smooth.grenander.test(
  y,
  delta,
  n.boot = 1000L,
  tau.c = NULL,
  skm.bw.multi = 0.7
)
}
\arguments{
\item{y}{A numeric vector of the observed survival times, \eqn{Y_i}.}

\item{delta}{A numeric vector of the censoring indicators, \eqn{\Delta_i}.}

\item{n.boot}{An integer of number of bootstrap samples.}

\item{tau.c}{A numeric of the end of the study time.
\code{NULL} indicates the maximum observed survival time \code{max(y)} is used.}

\item{skm.bw.multi}{A numeric of scaling to the bandwidth, \eqn{h_0}, for smoothing
the LCM of the KME.}
}
\value{
A list containing the results including the estimate
\eqn{\hat{f}_{nh}^{SG}(\tau_{G})} and
the bootstrap samples \eqn{\hat{f}_{nh}^{SG^\ast}(\tau_{G})}.
}
\description{
Testing the hypotheses:
\deqn{
   \tilde{H}_{0}: q_{1 - \epsilon} \geq \tau_{G}
   \quad\text{versus}\quad
   \tilde{H}_{a}: q_{1 - \epsilon} < \tau_{G},
}
using the smoothed Grenander estimator, \eqn{\hat{f}_{nh}^{SG}}.
}
\keyword{internal}
