% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gren.R
\name{smooth.kme.middle}
\alias{smooth.kme.middle}
\title{Kenrnel smoothing of the LCM of the KME (no boundary correction)}
\usage{
smooth.kme.middle(gcmlcm.fit, tau.c, x.grid.middle, bw)
}
\arguments{
\item{gcmlcm.fit}{A list of LCM of KME returned from \code{\link[fdrtool]{gcmlcm}}.}

\item{tau.c}{A numeric of the right endpoint of the support.}

\item{x.grid.middle}{A numeric vector of the points for which the value of Smoothed LCM of KME is computed.}

\item{bw}{A numeric of bandwidth}
}
\value{
A numeric vector of the computed value at \code{x.grid.middle}.
}
\description{
Kenrnel smoothing of the LCM of the KME (no boundary correction)
}
\keyword{internal}
