% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gren.R
\name{smooth.kme}
\alias{smooth.kme}
\title{Kenrnel smoothing of the LCM of the KME}
\usage{
smooth.kme(gcmlcm.fit, tau.c, n, skm.bw.multi = 0.7, grid.size = 5000L)
}
\arguments{
\item{gcmlcm.fit}{A list of LCM of KME returned from \code{\link[fdrtool]{gcmlcm}}.}

\item{tau.c}{A numeric of the end of the study time.}

\item{n}{A numeric of the sample size to determine the bandwidth.}

\item{skm.bw.multi}{A numeric of scaling to the bandwidth for smoothing
the LCM of the KME.}

\item{grid.size}{An integer of the grid size to generate a grid of points
for which the value of Smoothed LCM of KME is computed.}
}
\value{
A \code{data.frame}.
}
\description{
Kenrnel smoothing of the LCM of the KME:
\deqn{
 \int\frac{1}{h_{0}}k_{B,t}({\frac{t-u}{h_{0}}})\hat{F}_{n}^{G}(u)du,
}
where \eqn{k_{B,t}} is the boundary kernel, \eqn{h_{0}} is the bandwidth,
\eqn{\hat{F}_{n}^{G}} is the LCM of KME.
}
\keyword{internal}
