% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{tail.approx}
\alias{tail.approx}
\title{Lower bound of the density at \eqn{q_{1-\epsilon}}}
\usage{
\method{tail}{approx}(p.hat, tau, tau.c, eps = 0.05)
}
\arguments{
\item{p.hat}{A numeric of the estimated cure fraction \eqn{p} from KME.}

\item{tau}{A numeric of the parameter \eqn{\tau} for the test.}

\item{tau.c}{A numeric of the end of the study time.}

\item{eps}{A numeric of the \eqn{\epsilon} for the test.}
}
\value{
A numeric of the lower bound
}
\description{
The lower bound of \eqn{f(q_{1-\epsilon})}:
\deqn{{f(q_{1-\epsilon}) \geq
       \frac{(\epsilon - \eta)p}{\tau - \tau_{G}} \approx
       \frac{\epsilon p}{\tau - \tau_{G}}}.}
}
\keyword{internal}
